services/users/requirements.txt
coverage==4.5.1
flake8==3.5.0
flask-bcrypt==0.7.1
flask-migrate==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
Flask==1.0.2
psycopg2-binary==2.7.4
pyjwt==1.6.4
